docs(java): inspections to complete native IJ save actions
There’s stuff that used to be better handled when using the third-party plugin. And it’s weird to get yelled at for informative comments in if-else chains, especially when Sonar encourages us to end such chains with an extra else, even if it’s just to say “normal case, do nothing” or whatever.
Maybe this stuff about inspections could be placed in a dedicated section, I dunno. I doubt many people use a third-party save action thingy nowadays anyway.